// ui/src/types/enums.ts
export enum AccountType {
  SAVINGS = 'SAVINGS',
  CHECKING = 'CHECKING',
  CREDIT_CARD = 'CREDIT_CARD',
  FIXED_DEPOSIT = 'FIXED_DEPOSIT',
  CASH = 'CASH',
}

export enum CategoryType {
  EXPENSE = 'EXPENSE',
  INCOME = 'INCOME',
}

export enum TransactionType {
  EXPENSE = 'EXPENSE',
  INCOME = 'INCOME',
  TRANSFER = 'TRANSFER',
}